Classes page styling
