+++ client/src/styles/_base.scss
@@ -58,8 +58,6 @@ body {
   // The main content pane of the app
   #content {
     flex-grow: 1;
-    display: flex;
-    flex-direction: column;
   }
 }
