+++ client/src/styles/partials/_classes.scss
@@ -10,32 +10,15 @@
 }
 
 .classes-sidebar {
-  width: 70px;
   padding: 15px 10px;
-  &.phone {
-    width: 55px;
-    padding: 15px 5px;
-  }
-
-  display: flex;
-  flex-direction: column;
-  align-items: flex-end;
-  gap: 10px;
-
   &.expanded { 
     width: min(275px, 80vw);
   }
-
-  background-color: var(--bg-primary);
-
   height: 100vh;
 
   &.phone {
     height: calc(100vh - 64px);
   }
-
-  position: fixed;
-  right: 0px;
 }
 
 .classes-sidebar-bubble {
@@ -51,7 +34,6 @@
 
 .classes-sidebar-item {
   display: flex;
-  flex-direction: row;
   align-items: center;
   font-size: 1.2rem;
 
@@ -60,99 +42,17 @@
   }
 }
 
-.classes-burrito {
-  display: flex;
-  flex-direction: row;
-  max-width: 100vw;
-  height: 100%;
-
-  &.phone {
-    height: calc(100vh - 64px);
-  }
-}
-
-.classes-content {
-
-  padding: 10px calc(10px + 70px) 10px 10px;
-  &.phone {
-    padding: 10px calc(10px /* + 55px */) 10px 10px;
-  }
-
-  flex-grow: 1;
-  display: flex;
-  flex-direction: column;
-  scroll-behavior: smooth;
-
-  height: 100vh;
-
-  &.phone {
-    height: calc(100vh - 64px);
-  }
-
-  overflow-y: scroll;
-  overflow-x: hidden;
-}
-
-// Header
-.classes-header {
-  @include content-box($padding: 12px 20px, $bg-color: var(--active));
-  color: white;
-
-  // padding: 15px;
-  margin-bottom: 15px;
-  flex-direction: row;
-  align-items: center;
-}
-
 .classes-header-bubble {
   width: 40px;
   height: 40px;
-  border-radius: 40px;
-  flex-shrink: 0;
-  border: 2px solid var(--secondary);
-}
-
-.classes-header-text {
-  margin-left: 15px;
-  margin-bottom: 0;
 }
 
 // Nav Bar
 .classes-navbar {
-  display: flex;
-  flex-direction: row;
-  flex-wrap: wrap;
-  padding-right: 0;
-  padding-left: 0;
-  margin-bottom: 15px;
-  gap: 0.5rem;
-
   &-item {
     padding: 8px 12px;
     border-radius: 5px;
     font-size: 1.2rem;
-    display: flex;
-    align-items: center;
-    justify-content: center;
-    cursor: pointer;
     transition: all ease-in-out 100ms;
-    background-color: var(--content-secondary);
-
-    a {
-      color: var(--primary);
-      text-decoration: none;
-    }
-
-    // Selected nav items
-    &.selected {
-      background-color: var(--content-primary);
-      box-shadow: 0 10px 15px -3px rgba(0, 0, 0, 0.1), 0 4px 6px -2px rgba(0, 0, 0, 0.05);
-    }
   }
-}
-
-.classes-page {
-  padding-right: 0;
-  padding-left: 0;
-  height: 100%;
 }
